+++ src/core/components/loading/loading.component.tsx
