+++ src/core/components/sample/sample.component.tsx
