--- src/redux/actions/action.js
@@ -1,8 +1,0 @@
-const someAction = params => {
-  return {
-    type: "ACTION",
-    payload: params
-  }
-}
-
-export default someAction
