--- src/redux/actions/index.js
@@ -1,1 +1,0 @@
-export { default as action } from "./action"
