--- src/redux/initialState.js
@@ -1,5 +1,0 @@
-const initialState = {
-  // This will be add later
-}
-
-export default initialState
