+++ src/serviceWorker.tsx
@@ -39,7 +39,7 @@ export function register(config) {
         // Add some additional logging to localhost, pointing developers to the
         // service worker/PWA documentation.
         navigator.serviceWorker.ready.then(() => {
-          console.log("This web app is being served cache-first by a service " + "worker. To learn more, visit https://bit.ly/CRA-PWA")
+          console.log("This web app is being served cache-first by a service worker. To learn more, visit https://bit.ly/CRA-PWA")
         })
       } else {
         // Is not localhost. Just register service worker
@@ -64,7 +64,7 @@ function registerValidSW(swUrl, config) {
               // At this point, the updated precached content has been fetched,
               // but the previous service worker will still serve the older
               // content until all client tabs are closed.
-              console.log("New content is available and will be used when all " + "tabs for this page are closed. See https://bit.ly/CRA-PWA.")
+              console.log("New content is available and will be used when all tabs for this page are closed. See https://bit.ly/CRA-PWA.")
 
               // Execute callback
               if (config && config.onUpdate) {
